+++ lib/src/cookie_manager.dart
@@ -1,14 +1,21 @@
 import 'dart:async';
+import 'dart:io';
 
+import 'package:device_info/device_info.dart';
 import 'package:flutter/services.dart';
+import 'package:intl/intl.dart';
 
+import 'in_app_webview_controller.dart';
+import 'headless_in_app_webview.dart';
 import 'types.dart';
 
 ///Class that implements a singleton object (shared instance) which manages the cookies used by WebView instances.
 ///On Android, it is implemented using [CookieManager](https://developer.android.com/reference/android/webkit/CookieManager).
 ///On iOS, it is implemented using [WKHTTPCookieStore](https://developer.apple.com/documentation/webkit/wkhttpcookiestore).
 ///
-///**NOTE for iOS**: available from iOS 11.0+.
+///**NOTE for iOS below 11.0 (LIMITED SUPPORT!)**: in this case, almost all of the methods ([CookieManager.deleteAllCookies] is not supported!)
+///has been implemented using JavaScript because there is no other way to work with them on iOS below 11.0.
+///See https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies for JavaScript restrictions.
 class CookieManager {
   static CookieManager? _instance;
   static const MethodChannel _channel = const MethodChannel(
@@ -27,10 +34,17 @@ class CookieManager {
 
   static Future<dynamic> _handleMethod(MethodCall call) async {}
 
-  ///Sets a cookie for the given [url]. Any existing cookie with the same [host], [path] and [name] will be replaced with the new cookie. The cookie being set will be ignored if it is expired.
+  ///Sets a cookie for the given [url]. Any existing cookie with the same [host], [path] and [name] will be replaced with the new cookie.
+  ///The cookie being set will be ignored if it is expired.
   ///
   ///The default value of [path] is `"/"`.
   ///If [domain] is `null`, its default value will be the domain name of [url].
+  ///
+  ///[iosBelow11WebViewController] could be used if you need to set a session-only cookie using JavaScript (so [isHttpOnly] cannot be set, see: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
+  ///on the current URL of the [WebView] managed by that controller when you need to target iOS below 11. In this case the [url] parameter is ignored.
+  ///
+  ///**NOTE for iOS below 11.0**: If [iosBelow11WebViewController] is `null` or JavaScript is disabled for it, it will try to use a [HeadlessInAppWebView]
+  ///to set the cookie (session-only cookie won't work! In that case, you should set also [expiresDate] or [maxAge]).
   Future<void> setCookie(
       {required String url,
       required String name,
@@ -41,7 +55,8 @@ class CookieManager {
       int? maxAge,
       bool? isSecure,
       bool? isHttpOnly,
-      HTTPCookieSameSitePolicy? sameSite}) async {
+      HTTPCookieSameSitePolicy? sameSite,
+      InAppWebViewController? iosBelow11WebViewController}) async {
     if (domain == null) domain = _getDomainName(url);
 
     assert(url.isNotEmpty);
@@ -49,6 +64,56 @@ class CookieManager {
     assert(value.isNotEmpty);
     assert(domain.isNotEmpty);
     assert(path.isNotEmpty);
+
+    if (Platform.isIOS) {
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
+
+      var version = double.tryParse(iosInfo.systemVersion);
+
+      if (version != null && version < 11.0) {
+        var cookieValue = name + "=" + value + "; Domain=" + domain + "; Path=" + path;
+
+        if (expiresDate != null)
+          cookieValue += "; Expires=" + _getCookieExpirationDate(expiresDate);
+
+        if (maxAge != null)
+          cookieValue += "; Max-Age=" + maxAge.toString();
+
+        if (isSecure != null && isSecure)
+          cookieValue += "; Secure";
+
+        if (sameSite != null)
+          cookieValue += "; SameSite=" + sameSite.toValue();
+
+        cookieValue += ";";
+
+        if (iosBelow11WebViewController != null) {
+          InAppWebViewGroupOptions? options = await iosBelow11WebViewController.getOptions();
+          if (options != null && options.crossPlatform != null &&
+              options.crossPlatform!.javaScriptEnabled == true) {
+            await iosBelow11WebViewController.evaluateJavascript(
+                source: 'document.cookie="$cookieValue"');
+            return;
+          }
+        }
+
+        var setCookieCompleter = Completer<void>();
+        var headlessWebView = new HeadlessInAppWebView(
+          initialUrl: url,
+          onLoadStop: (controller, url) async {
+            await controller.evaluateJavascript(
+                source: 'document.cookie="$cookieValue"');
+            setCookieCompleter.complete();
+          },
+        );
+        await headlessWebView.run();
+        await setCookieCompleter.future;
+        await headlessWebView.dispose();
+
+        return;
+      }
+    }
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -66,8 +131,69 @@ class CookieManager {
   }
 
   ///Gets all the cookies for the given [url].
-  Future<List<Cookie>> getCookies({required String url}) async {
+  ///
+  ///[iosBelow11WebViewController] is used for getting the cookies (also session-only cookies) using JavaScript (cookies with `isHttpOnly` enabled cannot be found, see: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
+  ///from the current context of the [WebView] managed by that controller when you need to target iOS below 11. JavaScript must be enabled in order to work.
+  ///In this case the [url] parameter is ignored.
+  ///
+  ///**NOTE for iOS below 11.0**: All the cookies returned this way will have all the properties to `null` except for [Cookie.name] and [Cookie.value].
+  ///If [iosBelow11WebViewController] is `null` or JavaScript is disabled for it, it will try to use a [HeadlessInAppWebView]
+  ///to get the cookies (session-only cookies and cookies with `isHttpOnly` enabled won't be found!).
+  Future<List<Cookie>> getCookies({required String url, InAppWebViewController? iosBelow11WebViewController}) async {
     assert(url.isNotEmpty);
+
+    List<Cookie> cookies = [];
+
+    if (Platform.isIOS) {
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
+
+      var version = double.tryParse(iosInfo.systemVersion);
+
+      if (version != null && version < 11.0) {
+
+        if (iosBelow11WebViewController != null) {
+          InAppWebViewGroupOptions? options = await iosBelow11WebViewController.getOptions();
+          if (options != null && options.crossPlatform != null &&
+              options.crossPlatform!.javaScriptEnabled == true) {
+            List<String> documentCookies = (await iosBelow11WebViewController.evaluateJavascript(source: 'document.cookie') as String)
+                .split(';').map((documentCookie) => documentCookie.trim()).toList();
+            documentCookies.forEach((documentCookie) {
+              List<String> cookie = documentCookie.split('=');
+              cookies.add(Cookie(
+                name: cookie[0],
+                value: cookie[1],
+              )
+              );
+            });
+            return cookies;
+          }
+        }
+
+        var pageLoaded = Completer<void>();
+        var headlessWebView = new HeadlessInAppWebView(
+          initialUrl: url,
+          onLoadStop: (controller, url) async {
+            pageLoaded.complete();
+          },
+        );
+        await headlessWebView.run();
+        await pageLoaded.future;
+
+        List<String> documentCookies = (await headlessWebView.webViewController.evaluateJavascript(source: 'document.cookie') as String)
+            .split(';').map((documentCookie) => documentCookie.trim()).toList();
+        documentCookies.forEach((documentCookie) {
+          List<String> cookie = documentCookie.split('=');
+          cookies.add(Cookie(
+            name: cookie[0],
+            value: cookie[1],
+          )
+          );
+        });
+        await headlessWebView.dispose();
+        return cookies;
+      }
+    }
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -75,28 +201,86 @@ class CookieManager {
         await _channel.invokeMethod('getCookies', args);
     cookieListMap = cookieListMap.cast<Map<dynamic, dynamic>>();
 
-    List<Cookie> cookies = [];
-    for (var i = 0; i < cookieListMap.length; i++) {
+    cookieListMap.forEach((cookieMap) {
       cookies.add(Cookie(
-          name: cookieListMap[i]["name"],
-          value: cookieListMap[i]["value"],
-          expiresDate: cookieListMap[i]["expiresDate"],
-          isSessionOnly: cookieListMap[i]["isSessionOnly"],
-          domain: cookieListMap[i]["domain"],
+          name: cookieMap["name"],
+          value: cookieMap["value"],
+          expiresDate: cookieMap["expiresDate"],
+          isSessionOnly: cookieMap["isSessionOnly"],
+          domain: cookieMap["domain"],
           sameSite:
-              HTTPCookieSameSitePolicy.fromValue(cookieListMap[i]["sameSite"]),
-          isSecure: cookieListMap[i]["isSecure"],
-          isHttpOnly: cookieListMap[i]["isHttpOnly"],
-          path: cookieListMap[i]["path"]));
-    }
+          HTTPCookieSameSitePolicy.fromValue(cookieMap["sameSite"]),
+          isSecure: cookieMap["isSecure"],
+          isHttpOnly: cookieMap["isHttpOnly"],
+          path: cookieMap["path"]));
+    });
     return cookies;
   }
 
   ///Gets a cookie by its [name] for the given [url].
+  ///
+  ///[iosBelow11WebViewController] is used for getting the cookie (also session-only cookie) using JavaScript (cookie with `isHttpOnly` enabled cannot be found, see: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
+  ///from the current context of the [WebView] managed by that controller when you need to target iOS below 11. JavaScript must be enabled in order to work.
+  ///In this case the [url] parameter is ignored.
+  ///
+  ///**NOTE for iOS below 11.0**: All the cookies returned this way will have all the properties to `null` except for [Cookie.name] and [Cookie.value].
+  ///If [iosBelow11WebViewController] is `null` or JavaScript is disabled for it, it will try to use a [HeadlessInAppWebView]
+  ///to get the cookie (session-only cookie and cookie with `isHttpOnly` enabled won't be found!).
   Future<Cookie?> getCookie(
-      {required String url, required String name}) async {
+      {required String url, required String name, InAppWebViewController? iosBelow11WebViewController}) async {
+
     assert(url.isNotEmpty);
     assert(name.isNotEmpty);
+
+    if (Platform.isIOS) {
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
+
+      var version = double.tryParse(iosInfo.systemVersion);
+
+      if (version != null && version < 11.0) {
+
+        if (iosBelow11WebViewController != null) {
+          InAppWebViewGroupOptions? options = await iosBelow11WebViewController.getOptions();
+          if (options != null && options.crossPlatform != null &&
+              options.crossPlatform!.javaScriptEnabled == true) {
+            List<String> documentCookies = (await iosBelow11WebViewController.evaluateJavascript(source: 'document.cookie') as String)
+                .split(';').map((documentCookie) => documentCookie.trim()).toList();
+            for (var i = 0; i < documentCookies.length; i++) {
+              List<String> cookie = documentCookies[i].split('=');
+              if (cookie[0] == name)
+                return Cookie(
+                    name: cookie[0],
+                    value: cookie[1]);
+            }
+            return null;
+          }
+        }
+
+        var pageLoaded = Completer<void>();
+        var headlessWebView = new HeadlessInAppWebView(
+          initialUrl: url,
+          onLoadStop: (controller, url) async {
+            pageLoaded.complete();
+          },
+        );
+        await headlessWebView.run();
+        await pageLoaded.future;
+
+        List<String> documentCookies = (await headlessWebView.webViewController.evaluateJavascript(source: 'document.cookie') as String)
+            .split(';').map((documentCookie) => documentCookie.trim()).toList();
+        await headlessWebView.dispose();
+
+        for (var i = 0; i < documentCookies.length; i++) {
+          List<String> cookie = documentCookies[i].split('=');
+          if (cookie[0] == name)
+            return Cookie(
+                name: cookie[0],
+                value: cookie[1]);
+        }
+        return null;
+      }
+    }
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -123,18 +307,34 @@ class CookieManager {
   ///Removes a cookie by its [name] for the given [url], [domain] and [path].
   ///
   ///The default value of [path] is `"/"`.
-  ///If [domain] is `null` or empty, its default value will be the domain name of [url].
+  ///If [domain] is empty, its default value will be the domain name of [url].
+  ///
+  ///[iosBelow11WebViewController] is used for deleting the cookie (also session-only cookie) using JavaScript (cookie with `isHttpOnly` enabled cannot be deleted, see: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
+  ///from the current context of the [WebView] managed by that controller when you need to target iOS below 11. JavaScript must be enabled in order to work.
+  ///In this case the [url] parameter is ignored.
+  ///
+  ///**NOTE for iOS below 11.0**: If [iosBelow11WebViewController] is `null` or JavaScript is disabled for it, it will try to use a [HeadlessInAppWebView]
+  ///to delete the cookie (session-only cookie and cookie with `isHttpOnly` enabled won't be deleted!).
   Future<void> deleteCookie(
       {required String url,
       required String name,
       String domain = "",
-      String path = "/"}) async {
+      String path = "/",
+      InAppWebViewController? iosBelow11WebViewController}) async {
     if (domain.isEmpty) domain = _getDomainName(url);
 
     assert(url.isNotEmpty);
     assert(name.isNotEmpty);
-    assert(url.isNotEmpty);
-    assert(url.isNotEmpty);
+
+    if (Platform.isIOS) {
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
+      var version = double.tryParse(iosInfo.systemVersion);
+      if (version != null && version < 11.0) {
+        await setCookie(url: url, name: name, value: "", path: path, domain: domain, maxAge: -1, iosBelow11WebViewController: iosBelow11WebViewController);
+        return;
+      }
+    }
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -147,14 +347,33 @@ class CookieManager {
   ///Removes all cookies for the given [url], [domain] and [path].
   ///
   ///The default value of [path] is `"/"`.
-  ///If [domain] is `null` or empty, its default value will be the domain name of [url].
+  ///If [domain] is empty, its default value will be the domain name of [url].
+  ///
+  ///[iosBelow11WebViewController] is used for deleting the cookies (also session-only cookies) using JavaScript (cookies with `isHttpOnly` enabled cannot be deleted, see: https://developer.mozilla.org/en-US/docs/Web/HTTP/Cookies#restrict_access_to_cookies)
+  ///from the current context of the [WebView] managed by that controller when you need to target iOS below 11. JavaScript must be enabled in order to work.
+  ///In this case the [url] parameter is ignored.
+  ///
+  ///**NOTE for iOS below 11.0**: If [iosBelow11WebViewController] is `null` or JavaScript is disabled for it, it will try to use a [HeadlessInAppWebView]
+  ///to delete the cookies (session-only cookies and cookies with `isHttpOnly` enabled won't be deleted!).
   Future<void> deleteCookies(
-      {required String url, String domain = "", String path = "/"}) async {
+      {required String url, String domain = "", String path = "/",
+        InAppWebViewController? iosBelow11WebViewController}) async {
     if (domain.isEmpty) domain = _getDomainName(url);
 
     assert(url.isNotEmpty);
-    assert(url.isNotEmpty);
-    assert(url.isNotEmpty);
+
+    if (Platform.isIOS) {
+      DeviceInfoPlugin deviceInfo = DeviceInfoPlugin();
+      IosDeviceInfo iosInfo = await deviceInfo.iosInfo;
+      var version = double.tryParse(iosInfo.systemVersion);
+      if (version != null && version < 11.0) {
+        List<Cookie> cookies = await getCookies(url: url, iosBelow11WebViewController: iosBelow11WebViewController);
+        for (var i = 0; i < cookies.length; i++) {
+          await setCookie(url: url, name: cookies[i].name, value: "", path: path, domain: domain, maxAge: -1, iosBelow11WebViewController: iosBelow11WebViewController);
+        }
+        return;
+      }
+    }
 
     Map<String, dynamic> args = <String, dynamic>{};
     args.putIfAbsent('url', () => url);
@@ -164,6 +383,8 @@ class CookieManager {
   }
 
   ///Removes all cookies.
+  ///
+  ///**NOTE for iOS**: available from iOS 11.0+.
   Future<void> deleteAllCookies() async {
     Map<String, dynamic> args = <String, dynamic>{};
     await _channel.invokeMethod('deleteAllCookies', args);
@@ -175,5 +396,10 @@ class CookieManager {
     // ignore: unnecessary_null_comparison
     if (domain == null) return "";
     return domain.startsWith("www.") ? domain.substring(4) : domain;
+  }
+
+  String _getCookieExpirationDate(int expiresDate) {
+    var dateTime = DateTime.fromMillisecondsSinceEpoch(expiresDate).toUtc();
+    return DateFormat('EEE, d MMM yyyy hh:mm:ss', "en_US").format(dateTime) + ' GMT';
   }
 }